move TablePrefix enum to own file
