+++ nflow-engine/src/main/java/io/nflow/engine/internal/dao/TablePrefix.java
@@ -1,0 +1,16 @@
+package io.nflow.engine.internal.dao;
+
+public enum TablePrefix {
+  MAIN("nflow_"),
+  ARCHIVE("nflow_archive_");
+
+  private final String prefix;
+
+  TablePrefix(String prefix) {
+    this.prefix = prefix;
+  }
+
+  String nameOf(String name) {
+    return prefix + name;
+  }
+}
